flask-server/requirements.txt
aniso8601==9.0.1
attrs==23.2.0
blinker==1.7.0
click==8.1.7
flask==3.0.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
flask-restx==1.3.0
flask-sqlalchemy==3.1.1
gensim==4.3.2
greenlet==3.0.3
importlib-metadata==7.0.2
importlib-resources==6.3.1
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
numpy==1.24.4
pandas==2.0.3
pillow==10.2.0
pkgutil-resolve-name==1.3.10
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
referencing==0.34.0
rpds-py==0.18.0
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
smart-open==7.0.1
SQLAlchemy==2.0.28
threadpoolctl==3.3.0
typing-extensions==4.10.0
tzdata==2024.1
werkzeug==3.0.1
wrapt==1.16.0
zipp==3.18.1
